IDL: Transparent data forms for `service` and `func` (#920)
this proposes the “big” solution described in
https://github.com/dfinity-lab/motoko/pull/883; the “small” solution is
just too much of a hack (see
https://github.com/dfinity-lab/motoko/pull/913#issuecomment-557780787)

I think the idea is clear from this PR: We have two intro forms for
these types, one for references and one for the transparent public
actrors/methods, and serialize them differently, with a byte tag to
distinguish.

I am not sure if this is the best way of spec’ing this. Since this is
Andreas’s document, if we agree on the idea of this, I invite @rossberg
to take over the branch and reformulate as desired needed.