Spec: Opt decoding also from non-opt values
The problem that we had before this PR was the following:

* A sends values at type `record { foo : opt bool }`
  B recvs values at type `record { }`

  All izz well, as `record { foo : opt bool } <: record { }`

* Now `A` uses the rule
  ```
  not (null <: <datatype>)
  <datatype> <: <datatype'>
  -----------------------------
  <datatype> <: opt <datatype'>
  ```
  A sends values at type `record { foo : bool }` now

* Independently, `B` uses new rule about optional fields
  ```
  <datatype> <: <datatype'>
  -----------------------------
  <datatype> <: opt <datatype'>
  <nat> not in <fieldtype>;*
  record { <fieldtype>;* } <: record { <fieldtype'>;* }
  ------------------------------------------------------------------------------
  record { <fieldtype>;* } <: record { <nat> : opt <datatype'>; <fieldtype'>;* }
  ```
  B recvs values at type `record { foo : opt nat }`

Either of these steps on their own don’t break anything, because we
allow
 * `record { foo : bool } <: record { }`
 * `record { foo : opt bool } <: record { foo : opt nat }`

But both steps together are not allowed:

 * `record { foo : bool } <: record { foo : opt nat }`

The remidy seems to be to make the rule “when decoding an optional
field, be liberal in what you accept” even more liberal, and also allow
non-optional values.

This PR implements that.

The final rule
```
---------------------------------
<datatype> <: opt <datatype'>
```
looks very non-informative, but it seems to be the honest rule.

A direct generalization of the previous rule might have been
```
not (<datatype> <: opt <datatype'>)
---------------------------------
<datatype> <: opt <datatype'>
```
which is just odd.

An alternative might be to enumerate the cases, and have two rules
(which also matches the elaborated rules more closely):
```
not (<datatype> <: opt <datatype'>)
---------------------------------
opt <datatype> <: opt <datatype'>

not (null <: <datatype>)
not (<datatype> <: <datatype'>)
---------------------------------
opt <datatype> <: opt <datatype'>
```