derive candid type for functions (#119)
To export function types, an easy approach is to call `T::ty()` for each argument type. This works except for recursive types, and all the types are inlined in the service signature, which is not very human readable.

Another approach would be to implement an `export_type` function in `CandidType` trait to store type names and bindings to `TypeEnv`, but this doesn't work:
* Type names in different modules/namespaces collapse in Candid.
* Polymorphic types only know the instantiated types at the call site, and we need different names for each instantiation.
* Recursive types `Knot(TypeId)` need to convert to `Var(String)`.

This can all be solved via a global rewriting scheme. But then we can apply the rewriting to `T::ty()` directly without the need to call rewriting for each AST node. This is not ideal either, but good enough for now:
* We need a way to recover type names via `type_name::<T>()`, whose format is unspecified and long.
* Several Rust types can map to the same Candid type, and we only get to remember one name (currently we choose the shortest name). As a result, some of the type names in Rust is lost.
* Unless we do equivalence checking, recursive types can be unrolled and assigned to multiple names