Meta-Theory: Clarify transitive coherence
I tried to prove the “weak transitive coherence” that we claim in Coq:

    Theorem transitive_coherence:
      forall ta tb tc v1,
      ta <: tb ->
      tb <: tc ->
      v1 :: ta ->
      coerce tb tc (coerce ta tb v1) [=  coerce ta tc v1.

where [= allows more null on the left than on the right. I believed this
holds, but the proof doesn't go through.

A counter example is `bool <: opt bool <: opt opt bool`.

Coercing `true` in two steps goes via `opt true` to `opt opt true`.
Coercing directly goes to `null`, because the “constituent-to-opt” rule
`t <: opt t'` requires that `t'` is a non-opt type.

We added that restriction in 30f719fadbaccac for the reasons discussed
in https://github.com/dfinity/candid/issues/135#issuecomment-728833245

This PR just updates the prose to not claim wrong things.

(This is a good humbling reminder about how easy it is to go wrong when
one does not do formal proofs.)