fix typescript binding for references (#184)
Function/service types has two forms: 1) as a value, it's a principal id; 2) as a type signature, it's a function/service interface.

In typescript binding, we need to distinguish the two cases: Only show type signatures for the top-level service/function definitions. The input/output arguments are always of principal type.