Spec: Clarify that the type table may only contain composite types
A developers was confused about which types types need to go behind the type
table, and which don’t, partly because the Spec uses “primitive” to mean
primitive data in one place, and non-composite in others
https://forum.dfinity.org/t/403-error-when-i-call-the-management-canister/7619/14?u=nomeata

This is mostly history: We had that wording before we introduced the
`principal` type.

This clarifies this bit by avoiding the term “primitive” here, and
explicitly pointing out that the type table is only for composite types
(A-normal form).

One could also re-label `principal` as a `<primtype>` to get the same
effect. I did not propose that expecting resistance, but wouldn’t mind
that fix either.