feat: derive CandidType for Rc and Arc
It is often desirable to use reference-counting in canisters for large
blobs to avoid expensive copies.  Unfortunately, this optimization
requires defining newtypes and implementing CandidType for them
(example:
https://github.com/dfinity/certified-assets/commit/47804eb70f44d2e5c73da26f0009540330293eb2).

This change adds impls for Rc and Arc to the candid library to make
this kind of optimizations less painful to implement.