feat(didc-js): encode and validate specific types and service args (#602)
**Overview**
This PR extends the `didc-js` wasm module with capability to encode
values of specific types, and values for the service arguments, with the
purpose of also validating them against those types. This is useful for
JS clients where the user is expected to produce valid candid value for
a particular type.

**Requirements**
- the `encode` function can receive type information other than
`service_method` such as a type name or instruction to use the service
argument types

**Considered Solutions**
I did not find any other solution.

**Recommended Solution**
Since `encode` already has this functionality to encode (and validate)
service method arguments, this seems like a good place to extend the
functionality.

**Considerations**