refactor!: move `IDL*` types into the candid crate (#622)
**Overview**
Move the parser's `IDL*` types to the `candid` crate, in order to share
it between both the `candid` and the `candid_parser` crates.

The main reason is that we want to use the `IDL*` types in the bindings
generators, instead of bringing around the `TypeEnv`, which holds inner
`Type`s that are not meant for codegen.

**Considerations**
Introduces breaking changes.