feat: support doc comments in Motoko bindgen (#634)
**Overview**
Adds support for reflecting doc comments from Candid declaration files
to the Motoko generated bindings.

Internally, it adds the `IDLMergedProg` which is constructed when
parsing the Candid declarations and is passed to the Motoko bindings
generator. It containts the syntax types along with the associated
comments, and it's traversed in parallel with the inner types by the
binding generator.

Follow-up PRs will add support for doc comments in the other bindings
generators.

Closes #625.

---------

Co-authored-by: ilbertt <luca.bertelli@dfinity.org>