feat: JavaScript target IDL type object exports and factory function deprecations (#665)
**Overview**

The JavaScript target would be more useful (especially to Azle
developers) if it exported the IDL type objects that it generates. We
also no longer need the factory functions as of `3.0.0`, thus we will
mark them deprecated and provide direct exports for that previous
functionality.

**Requirements**

The JavaScript target should export all of the IDL type objects,
including the generated `IDL.Service` and `init args`, without the need
to use a factory function.

**Considerations**

There are no breaking changes besides some formatting of the generated
JavaScript files. The `idflFactory` and `init` functions have been
marked as deprecated but not removed. Developers can choose to import
the `IDL` type objects now, but they are not forced to. This is mainly
an ergonomic change with deprecations to prepare for a future with the
best developer experience by default.