feat!: collect names for return types in parser (#684)
**Overview**
Named arguments can also be in the return type according to the spec.
This PR updates the grammar to collect the names of the return types.

As a consequence, it fixes a regression introduced in
https://github.com/dfinity/candid/pull/612/files#diff-c5ff9610c5d6538ec79f3d465965d2b3a3e3b628b16854a91b4b4a90ba766a57L220-L230,
in order to keep parsing the return types with arguments without errors.

Updates a test case to check if the behavior has been restored.